# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Smart Expense & Budget Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /budgets:
    get:
      operationId: listBudgets
      tags: [budgets]
      summary: List user budgets
      responses:
        "200":
          description: List of budgets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BudgetWithSpent"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createBudget
      tags: [budgets]
      summary: Create a new budget
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBudgetBody"
      responses:
        "201":
          description: Budget created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /budgets/{id}:
    get:
      operationId: getBudget
      tags: [budgets]
      summary: Get a budget by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Budget details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BudgetWithSpent"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateBudget
      tags: [budgets]
      summary: Update a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBudgetBody"
      responses:
        "200":
          description: Updated budget
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Budget"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteBudget
      tags: [budgets]
      summary: Delete a budget
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Budget deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Budget:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        limit:
          type: number
        period:
          type: string
          enum: [monthly, weekly, yearly]
        startDate:
          type: string
          format: date-time
        endDate:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - name
        - limit
        - period
        - startDate
        - createdAt
    BudgetWithSpent:
      allOf:
        - $ref: "#/components/schemas/Budget"
        - type: object
          properties:
            spent:
              type: number
            remaining:
              type: number
            percentUsed:
              type: number
          required:
            - spent
            - remaining
            - percentUsed
    CreateBudgetBody:
      type: object
      properties:
        name:
          type: string
        limit:
          type: number
        period:
          type: string
          enum: [monthly, weekly, yearly]
        startDate:
          type: string
          format: date-time
        endDate:
          type: ["string", "null"]
      required:
        - name
        - limit
        - period
        - startDate
    UpdateBudgetBody:
      type: object
      properties:
        name:
          type: string
        limit:
          type: number
        period:
          type: string
          enum: [monthly, weekly, yearly]
        endDate:
          type: ["string", "null"]
...